change ints to floats (prior to testing)

+++ pybricks/tools/__stub/__tools.py
@@ -1,8 +1,8 @@
-def wait(time: int):
+def wait(time: float):
     """
     Pauses the user program for a specified amount of time.
 
     Args:
-        time (int): How long to wait in milliseconds.
+        time (float): How long to wait in milliseconds.
     """
     ...
